WIP: __array_ufunc__ for multiarray functions
